# db: fix levelIter TrySeekUsingNext optimization