The optimization added in https://github.com/cockroachdb/pebble/pull/2058 began using an enabled TrySeekUsingNext
flag to avoid re-seeking within a level's file metadata. This
optimization was dependent on the invariant that the iterator remained
positioned at the previous seek key, so that a subsequent seek to a
larger key does not need to backtrack.

This invariant wasn't strictly preserved by the levelIter during
SeekPrefixGE calls. During a SeekPrefixGE, the sstable iterator may
return nil despite the existence of sstable keys greater than the seek
key if the sstable's bloom filter excludes the seek prefix. If the
sstable DOES NOT contain any range tombstones, the levelIter does not
advance to the next file if the file's largest bound has a prefix larger
than the seek prefix, returning nil, else it does advance since the next
file could contain the seek prefix.

However, if the file DOES contain range tombstones, the levelIter
returns a synthetic largest boundary key so that the file remains open
until the merging iterator passes beyond its bounds. This ensures the
file's range deletions' effects on other keys are observed. If another
level returned a key greater than this largest boundary key (eg, because
the other level doesn't restrict results to the seek prefix), the
merging iterator could step beyond the level's synthetic boundary key.
This step could advance the levelIter to the next file, despite its
irrelevance to the current prefix. This step would also break the
invariant that the level iterator remained positioned at the seek key.

This commit fixes the bug by comparing the synthetic boundary key to the
seek prefix, avoiding ever Next-ing the level iterator beyond the seek
prefix.

Based on our current understanding of the issue, this commit will also
fix https://github.com/cockroachdb/cockroach/issues/89327 by preventing advancement of the
levelIter beyond the seek prefix. See the related new microbenchmark.

```
name                     old time/op  new time/op  delta
SeekPrefixTombstones-16   706µs ± 2%     1µs ± 5%  -99.92%  (p=0.000 n=8+10)
```

Fix https://github.com/cockroachdb/pebble/issues/2084.